law-predictor/backend/requirements.txt
fastapi
uvicorn[standard]
faiss-cpu
sentence-transformers
transformers
torch
ollama
pydantic
requests
beautifulsoup4
pdfplumber
tqdm
scikit-learn
python-multipart
joblib